pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["Sameer Segal <sameersegal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gradio = "^4.19.1"
openai = "^1.12.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
